+++ src/pages/about.js
@@ -1,7 +1,7 @@
 import React from "react"
 import { Link as RouterLink } from "gatsby"
 
-import { PageLayout } from "../components/PageLayout/PageLayout"
+import { PageLayout } from "../layouts/PageLayout/PageLayout"
 import SEO from "../components/SEO"
 import SayHi from "../components/SayHi"
