+++ src/pages/writings.js
@@ -1,7 +1,7 @@
 import React from "react"
 import { graphql } from "gatsby"
 
-import { PageLayout } from "../components/PageLayout/PageLayout"
+import { PageLayout } from "../layouts/PageLayout/PageLayout"
 import SEO from "../components/SEO"
 import Link from "../components/Link"
